% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_fourPlot}
\alias{eda_fourPlot}
\title{EDA Assumptions Testing - Four Plot}
\usage{
eda_fourPlot(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\description{
This function takes in a vector of numerical data and creates a 4-Plot
consisting of:
1. Run Sequence Plot
2. Lag 1 Plot
3. Histogram / Density Plot
4. QQ Normality Plot
}
\examples{
eda_fourPlot(rnorm(1000))
}
