% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_locationDrift}
\alias{eda_locationDrift}
\title{EDA Assumptions Testing - Drift in Location}
\usage{
eda_locationDrift(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\description{
This function takes in a vector of numerical data and tests drift in location using
a simple regression model
}
\examples{
eda_locationDrift(rnorm(1000))
}
