% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_normality}
\alias{eda_normality}
\title{EDA Assumptions Testing - Normality}
\usage{
eda_normality(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\description{
This function takes in a vector of numerical data and tests if the data
came from a normal distribution.
}
\examples{
eda_normality(rnorm(1000))
}
