% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_randomness}
\alias{eda_randomness}
\title{EDA Assumptions Testing - Randomness}
\usage{
eda_randomness(y, plot.it = TRUE)
}
\arguments{
\item{y}{Numeric vector.}

\item{plot.it}{Boolean value that controls whether autocorrelation / partial autocorrelation plots are created. Defaults to TRUE.}
}
\description{
This function takes in a vector of numerical data and tests for patterns in the data.
}
\examples{
eda_randomness(rnorm(1000))
eda_randomness(rnorm(1000), plot.it = FALSE)
}
