% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_SummaryStats}
\alias{eda_SummaryStats}
\title{Summary Statistics}
\usage{
eda_SummaryStats(y)
}
\arguments{
\item{y}{Numeric vector.}
}
\description{
This function takes in a vector of numerical data and computes
summary statistics
}
\examples{
eda_SummaryStats(rnorm(1000))
}
