% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_varianceDrift}
\alias{eda_varianceDrift}
\title{EDA Assumptions Testing - Drift in Variation}
\usage{
eda_varianceDrift(y, testType = "barlett")
}
\arguments{
\item{y}{Numeric vector.}

\item{testType}{Variance Homogeniety Test to use.  Defaults to "barlett", other option is "levene".}
}
\description{
This function takes in a vector of numerical data and tests homoscedasticity
}
\examples{
eda_varianceDrift(rnorm(1000))
eda_varianceDrift(runif(1000), testType = "levene")
}
